add e nums
