+++ prisma/schema.prisma
@@ -1,3 +1,4 @@
+// Use https://prisma-editor.vercel.app/schema/ to edit diagram
 generator client {
   provider = "prisma-client-js"
 }
@@ -6,20 +7,6 @@ datasource db {
   provider = "mysql"
   url      = env("DATABASE_URL")
 }
-
-// ===================== Database Versioning Model =====================
-
-/// Tracks the versions of the database for major migrations or updates.
-model DatabaseVersion {
-  id            String   @id @default(uuid())
-  version       String   @unique   
-  description   String?           
-  applied_at    DateTime @default(now())
-  successful    Boolean  @default(true)  
-  createdAt     DateTime @default(now())
-  updatedAt     DateTime @updatedAt
-}
-
 
 // ===================== Enums =====================
 
@@ -45,6 +32,82 @@ enum ProjectStatusEnum {
   DEPLOYED
   STARTUP
 }
+
+enum ProjectDomainEnum {
+  AI
+  ML
+  AR_VR
+  BLOCKCHAIN
+  IOT
+  HEALTHTECH
+  FINTECH
+  EDTECH
+  AGRITECH
+  ECOMMERCE
+  SOCIAL
+  GAMING
+  SECURITY
+  DATA_ANALYTICS
+  ENTERTAINMENT
+  SUSTAINABILITY
+}
+
+enum ProjectTypeEnum {
+  MOBILE
+  WEB
+  HARDWARE
+  DESKTOP
+  WEARABLE
+}
+
+enum SDGGoalEnum {
+  NO_POVERTY
+  ZERO_HUNGER
+  GOOD_HEALTH
+  QUALITY_EDUCATION
+  GENDER_EQUALITY
+  CLEAN_WATER
+  AFFORDABLE_ENERGY
+  DECENT_WORK
+  INDUSTRY_INNOVATION
+  REDUCED_INEQUALITIES
+  SUSTAINABLE_CITIES
+  RESPONSIBLE_CONSUMPTION
+  CLIMATE_ACTION
+  LIFE_BELOW_WATER
+  LIFE_ON_LAND
+  PEACE_JUSTICE
+  PARTNERSHIPS
+}
+
+enum TechStackEnum {
+  REACT
+  ANGULAR
+  VUE
+  NODE
+  PYTHON
+  DJANGO
+  FLASK
+  JAVA
+  SPRING
+  DOTNET
+  PHP
+  LARAVEL
+  ANDROID
+  IOS
+  FLUTTER
+  REACT_NATIVE
+  FIREBASE
+  AWS
+  MONGODB
+  MYSQL
+  POSTGRESQL
+  TENSORFLOW
+  PYTORCH
+  ARDUINO
+  RASPBERRY_PI
+}
+
 
 // ===================== Models =====================
 
@@ -118,14 +181,21 @@ model ProjectStatus {
 model ProjectAssociation {
   id          String           @id @default(uuid())
   project_id  String
-  type        AssociationType  // Specifies the type of association
-  value       String           // The actual value (e.g., name of tech, domain, etc.)
-  project     ProjectContent   @relation(fields: [project_id], references: [content_id], onDelete: Restrict)
-  createdAt   DateTime         @default(now())
-  updatedAt   DateTime         @updatedAt
+  type        AssociationType  // Can stay the same, if it refers to the enums
+  value       String            // This should be replaced by enums based on type
+
+  domain      ProjectDomainEnum?  
+  projectType ProjectTypeEnum?  
+  sdgGoal     SDGGoalEnum?  
+  techStack   TechStackEnum?  
+
+  project     ProjectContent    @relation(fields: [project_id], references: [content_id], onDelete: Restrict)
+  createdAt   DateTime          @default(now())
+  updatedAt   DateTime          @updatedAt
 
   @@index([project_id], name: "project_association_index")
 }
+
 
 // Stores individual slides related to a project.
 model ProjectSlide {
